crypto-analytics/requirements.txt
flask
gunicorn
yfinance
pandas
numpy
scipy
scikit-learn
joblib
requests
tqdm
python-dotenv
aiohttp
flask-sqlalchemy
flask-cors
bcrypt
stripe
openai
ccxt
matplotlib
fredapi
pytest
pytest-flask
